feat: AWS SNS Client 설정
